Add background request support via PendingIntent
